+++ src/lib/zt/if-condition.ts
@@ -1,21 +1,33 @@
 
-import {
-    type IRenderableKargs,
-    type IRenderableOutput,
-    type IRenderable
-} from "../core/renderable";
-import {
-    type TagIdentity,
-    tagIdentity
-} from "./identity";
+import { type IRenderable } from "../core/renderable";
+import { tagIdentity } from "./identity";
 
-export const ifCondition = <
+/**
+ * Conditionaly return given renderable
+ * 
+ * `zt.if` returns `zt.empty` when condition is not met
+ */
+export function ifCondition<
     T extends IRenderable<any, any>
->(condition: unknown, template: T) => {
-    return (!!condition ? template : tagIdentity) as (
-        IRenderable<
-            IRenderableKargs<T> | IRenderableKargs<TagIdentity>,
-            IRenderableOutput<T> | IRenderableOutput<TagIdentity>
-        >
-    )
+>(
+    condition: false | 0 | "" | null | undefined,
+    template: T,
+): IRenderable<void, []>
+
+/**
+ * Conditionaly return given renderable
+ * 
+ * `zt.if` returns `zt.empty` when condition is not met
+ */
+export function ifCondition<
+    T extends IRenderable<any, any>,
+>(condition: any, template: T): T
+
+/**
+ * `zt.if` implementation
+ */
+export function ifCondition<
+    T extends IRenderable<any, any>,
+>(condition: any, template: T) {
+    return (condition ? template : tagIdentity) as any
 }